[xla] Extract a separate collective_rendezvous library
