+++ xla/backends/cpu/collectives/BUILD
@@ -120,7 +120,7 @@ cc_library(
         "//xla/core/collectives:collectives_registry",
         "//xla/core/collectives:communicator",
         "//xla/core/collectives:rank_id",
-        "//xla/service:collective_ops_utils",
+        "//xla/service:collective_rendezvous",
         "@com_google_absl//absl/base",
         "@com_google_absl//absl/log",
         "@com_google_absl//absl/log:check",
@@ -163,7 +163,8 @@ cc_library(
         "//xla:xla_data_proto_cc",
         "//xla/core/collectives:communicator",
         "//xla/core/collectives:rank_id",
-        "//xla/service:collective_ops_utils",
+        "//xla/core/collectives:reduction_kind",
+        "//xla/service:collective_rendezvous",
         "//xla/service:rendezvous",
         "//xla/stream_executor:device_address",
         "//xla/tsl/lib/math:math_util",
@@ -249,7 +250,7 @@ xla_cc_test(
         "//xla/pjrt/distributed:in_memory_key_value_store",
         "//xla/pjrt/distributed:key_value_store_interface",
         "//xla/runtime:device_id",
-        "//xla/service:collective_ops_utils",
+        "//xla/service:collective_rendezvous",
         "//xla/stream_executor:device_address",
         "//xla/tsl/concurrency:async_value",
         "//xla/tsl/lib/core:status_test_util",
