+++ xla/backends/cpu/collectives/cpu_collectives.cc
@@ -23,7 +23,7 @@ limitations under the License.
 #include "xla/core/collectives/collectives.h"
 #include "xla/core/collectives/collectives_registry.h"
 #include "xla/core/collectives/communicator.h"
-#include "xla/service/collective_ops_utils.h"
+#include "xla/service/collective_rendezvous.h"
 #include "xla/util.h"
 #include "tsl/platform/casts.h"
