+++ xla/backends/cpu/runtime/BUILD
@@ -386,7 +386,7 @@ cc_library(
         "//xla:shape_util",
         "//xla/backends/cpu/collectives:cpu_collectives",
         "//xla/core/collectives:communicator",
-        "//xla/service:collective_ops_utils",
+        "//xla/service:collective_rendezvous",
         "//xla/tsl/concurrency:async_value",
         "//xla/tsl/platform:statusor",
         "@com_google_absl//absl/container:inlined_vector",
@@ -513,8 +513,9 @@ cc_library(
         "//xla:util",
         "//xla/backends/cpu/collectives:cpu_collectives",
         "//xla/core/collectives:communicator",
+        "//xla/core/collectives:reduction_kind",
         "//xla/service:buffer_assignment",
-        "//xla/service:collective_ops_utils",
+        "//xla/service:collective_rendezvous",
         "//xla/tsl/concurrency:async_value",
         "//xla/tsl/platform:statusor",
         "@com_google_absl//absl/container:inlined_vector",
@@ -540,7 +541,7 @@ cc_library(
         "//xla/backends/cpu/collectives:cpu_collectives",
         "//xla/core/collectives:communicator",
         "//xla/service:buffer_assignment",
-        "//xla/service:collective_ops_utils",
+        "//xla/service:collective_rendezvous",
         "//xla/tsl/concurrency:async_value",
         "//xla/tsl/platform:logging",
         "//xla/tsl/platform:statusor",
@@ -565,7 +566,8 @@ cc_library(
         "//xla:xla_data_proto_cc",
         "//xla/backends/cpu/collectives:cpu_collectives",
         "//xla/core/collectives:communicator",
-        "//xla/service:collective_ops_utils",
+        "//xla/core/collectives:reduction_kind",
+        "//xla/service:collective_rendezvous",
         "//xla/tsl/concurrency:async_value",
         "//xla/tsl/platform:statusor",
         "@com_google_absl//absl/container:inlined_vector",
@@ -592,7 +594,7 @@ cc_library(
         "//xla/core/collectives:communicator",
         "//xla/core/collectives:rank_id",
         "//xla/service:buffer_assignment",
-        "//xla/service:collective_ops_utils",
+        "//xla/service:collective_rendezvous",
         "//xla/service:computation_placer",
         "//xla/tsl/concurrency:async_value",
         "//xla/tsl/platform:statusor",
@@ -630,6 +632,7 @@ cc_library(
         "//xla/runtime:resource_use",
         "//xla/service:buffer_assignment",
         "//xla/service:collective_ops_utils",
+        "//xla/service:collective_rendezvous",
         "//xla/service:computation_placer",
         "//xla/service:hlo_proto_cc",
         "//xla/stream_executor:device_address",
