+++ xla/backends/cpu/runtime/all_reduce_thunk.h
@@ -20,8 +20,8 @@ limitations under the License.
 
 #include "absl/status/statusor.h"
 #include "xla/backends/cpu/runtime/collective_thunk.h"
+#include "xla/core/collectives/reduction_kind.h"
 #include "xla/service/buffer_assignment.h"
-#include "xla/service/collective_ops_utils.h"
 #include "xla/tsl/concurrency/async_value_ref.h"
 
 namespace xla::cpu {
