+++ xla/backends/cpu/runtime/all_to_all_thunk.cc
@@ -29,7 +29,7 @@ limitations under the License.
 #include "xla/core/collectives/communicator.h"
 #include "xla/future.h"
 #include "xla/service/buffer_assignment.h"
-#include "xla/service/collective_ops_utils.h"
+#include "xla/service/collective_rendezvous.h"
 #include "xla/shape.h"
 #include "xla/shape_util.h"
 #include "xla/tsl/concurrency/async_value_ref.h"
