+++ xla/backends/cpu/runtime/collective_permute_thunk.cc
@@ -38,7 +38,7 @@ limitations under the License.
 #include "xla/core/collectives/rank_id.h"
 #include "xla/future.h"
 #include "xla/service/buffer_assignment.h"
-#include "xla/service/collective_ops_utils.h"
+#include "xla/service/collective_rendezvous.h"
 #include "xla/service/computation_placer.h"
 #include "xla/shape.h"
 #include "xla/shape_util.h"
@@ -75,9 +75,10 @@ CollectivePermuteThunk::Execute(const ExecuteParams& params) {
   Thunk::CollectiveExecuteParams* collective_params = params.collective_params;
   TF_RET_CHECK(collective_params) << "Collectives parameters are not set";
 
-  ABSL_ASSIGN_OR_RETURN(DeviceAssignment::LogicalID logical_id,
-                   collective_params->device_assignment->LogicalIdForDevice(
-                       collective_params->global_device_id));
+  ABSL_ASSIGN_OR_RETURN(
+      DeviceAssignment::LogicalID logical_id,
+      collective_params->device_assignment->LogicalIdForDevice(
+          collective_params->global_device_id));
 
   int32_t logical_device_id = op_params().has_channel_id
                                   ? logical_id.computation_id
