+++ xla/backends/cpu/runtime/collective_thunk.h
@@ -35,7 +35,7 @@ limitations under the License.
 #include "xla/runtime/device_id.h"
 #include "xla/runtime/resource_use.h"
 #include "xla/service/buffer_assignment.h"
-#include "xla/service/collective_ops_utils.h"
+#include "xla/service/collective_rendezvous.h"
 #include "xla/shape.h"
 #include "xla/stream_executor/device_address.h"
 #include "xla/tsl/concurrency/async_value_ref.h"
