+++ xla/backends/cpu/runtime/reduce_scatter_thunk.h
@@ -20,7 +20,7 @@ limitations under the License.
 
 #include "absl/status/statusor.h"
 #include "xla/backends/cpu/runtime/collective_thunk.h"
-#include "xla/service/collective_ops_utils.h"
+#include "xla/core/collectives/reduction_kind.h"
 #include "xla/tsl/concurrency/async_value_ref.h"
 #include "xla/xla_data.pb.h"
