+++ xla/service/collective_rendezvous.h
@@ -1,0 +1,91 @@
+/* Copyright 2026 The OpenXLA Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+
+#ifndef XLA_SERVICE_COLLECTIVE_RENDEZVOUS_H_
+#define XLA_SERVICE_COLLECTIVE_RENDEZVOUS_H_
+
+#include <cstdint>
+#include <string>
+#include <utility>
+#include <vector>
+
+#include "absl/strings/string_view.h"
+#include "xla/executable_run_options.h"
+#include "xla/runtime/device_id.h"
+
+namespace xla {
+
+// Key that identifies a particular Rendezvous object in our global hashtable.
+// This determines which calls to ExecuteOnStream communicate with each other.
+// The rules are as follows.
+//
+// * Only ops with the same RunId can communicate with each other. (This is the
+//   whole purpose of RunId).
+//
+// * Only ops with the same set of participating replicas can communicate with
+//   each other.  This is how we separate out different replica groups (e.g. a
+//   single AllReduce HLO might do two reductions, between say GPUs {0,2} and
+//   {1,3}).
+//
+// * Only ops with the same opcode can communicate with each other.  At the
+//   moment we only support kAllReduce, so we don't check for this explicitly.
+//
+// * For cross-module all-reduces (i.e. instr->channel_id().has_value()),
+//   only ops with the same value for channel_id() can communicate with each
+//   other.
+//
+// * For cross-replica (i.e. same-module) all-reduces (i.e.
+//   !channel_id().has_value()), only ops from the same module (as
+//   identified by its unique_id()) can communicate with each other.
+//
+struct RendezvousKey {
+  enum CollectiveOpKind {
+    kCrossModule,
+    kCrossReplica,
+  };
+
+  explicit RendezvousKey(const RunId& run_id,
+                         std::vector<GlobalDeviceId> global_devices,
+                         int num_local_participants,
+                         CollectiveOpKind collective_op_kind, int64_t op_id);
+
+  template <typename H>
+  friend H AbslHashValue(H h, const RendezvousKey& k) {
+    return H::combine(std::move(h), k.run_id, k.global_devices,
+                      k.num_local_participants, k.collective_op_kind, k.op_id);
+  }
+  friend bool operator==(const RendezvousKey& a, const RendezvousKey& b) {
+    return a.run_id == b.run_id && a.global_devices == b.global_devices &&
+           a.num_local_participants == b.num_local_participants &&
+           a.collective_op_kind == b.collective_op_kind &&  //
+           a.op_id == b.op_id;
+  }
+  friend bool operator!=(const RendezvousKey& a, const RendezvousKey& b) {
+    return !(a == b);
+  }
+
+  absl::string_view CollectiveOpKindString() const;
+  std::string ToString() const;
+
+  RunId run_id;
+  std::vector<GlobalDeviceId> global_devices;
+  int num_local_participants;
+  CollectiveOpKind collective_op_kind;
+  int64_t op_id;
+};
+
+}  // namespace xla
+
+#endif  // XLA_SERVICE_COLLECTIVE_RENDEZVOUS_H_
